Create Main
This file contains the bulk of the program's coding. While we call functions from separate included files (which I have named RGBalance.h and colourDictionary.h) this code applies the calculated average reading of each wavelength of red, green and blue light to an array for further maths and printing. In order to be the most accurate we can with this system the program will read the data from a sample of white, then a sample of black, to use as maximum and minimum values respectively. This function will only run once, afterwards the program will begin to identify colour wavelengths using these balancing values.