Build the Circuit (w/ picture guides!)
Materials:
- 2x 220 ohms resistor
- m/m jumper wires
- Cathode RGB LED
- Light-dependent resistor
- Breadboard
- Arduino

As per this setup, the LED is connected to the Arduino through pins 13, 12 and 11 with a resistor going out to ground. The LDR is connected to analog pin A5 with a voltage divider applied by use of a resistor connecting to ground and a supply of 5V from the Arduino.